move calculatePlayerPositions into PlayerCircles file. its not used anywhere else.
