+++ ui/screens/Game/components/CardTable/PlayerCircles.tsx
@@ -3,7 +3,7 @@ import { Container, Text } from "@/ui/elements";
 import { useCurrentGame } from "@/ui/screens/Game/PlayingContext";
 import React, { useEffect, useMemo } from "react";
 import { useCardTable } from "./context";
-import { calculatePlayerPositions } from "./utils";
+import { degreesToRadians, radiansToDegrees } from "./utils";
 
 export function PlayerCircles() {
   const { game } = useCurrentGame();
@@ -53,4 +53,29 @@ export function PlayerCircles() {
       })}
     </>
   );
+}
+
+export function calculatePlayerPositions(
+  players: string[],
+  radius: number,
+  currentUserId: string
+): { playerId: string; x: number; y: number; rotation: number }[] {
+  const currentUserIndex = players.findIndex(
+    (player) => player === currentUserId
+  );
+  const currentUserAngle = (currentUserIndex * 360) / players.length;
+  const rotationDegrees = 90 - currentUserAngle;
+
+  return players.map((player, index) => {
+    const baseAngle = (index * 2 * Math.PI) / players.length;
+    const rotatedAngle = baseAngle + degreesToRadians(rotationDegrees);
+    const x = radius * Math.cos(rotatedAngle);
+    const y = radius * Math.sin(rotatedAngle);
+    return {
+      x,
+      y,
+      rotation: radiansToDegrees(rotatedAngle) - 90,
+      playerId: player,
+    };
+  });
 }
